Move Multiplexed to its own module
